Support fractional shares for Polymarket orders
Polymarket uses fractional shares (e.g., 1.94 contracts after fees).
- Frontend: parseFloat for PM, Math.floor for KS (whole contracts only)
- api_server: store count as float, cast to int for Kalshi only

+++ api_server.py
@@ -710,7 +710,7 @@ async def websocket_trade(websocket: WebSocket):
                 platform = data.get("platform", "")
                 action = data.get("action", "")
                 side = data.get("side", "")
-                count = int(data.get("count", 0))
+                count = float(data.get("count", 0))
                 price = data.get("price")
                 order_type = data.get("order_type", "limit")
                 ticker = data.get("ticker", "")
@@ -763,7 +763,7 @@ async def websocket_trade(websocket: WebSocket):
                             result = await asyncio.to_thread(
                                 place_kalshi_order,
                                 order["ticker"], order["action"], order["side"],
-                                order["count"], ks_price, order["order_type"],
+                                int(order["count"]), ks_price, order["order_type"],
                             )
                         else:
                             from clients.executor import place_polymarket_order
